refactor: move useDistinctLabelMediaDevices to vera
