--- libs/ui/src/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useDistinctLabelMediaDevices } from './useDistinctLabelMediaDevices';
